refactor(csr): rename plasticity_binary module to plasticity_binary_csr
Rename _csr/plasticity_binary.py -> _csr/plasticity_binary_csr.py and its
test to plasticity_binary_csr_test.py, disambiguating it from the dense
counterpart (_dense/plasticity_binary.py) and aligning with the _csc naming
(plasticity_binary_csc.py). Update all intra-package imports (_csr/__init__.py,
_csr/main.py, plasticity_binary_csc.py) and the in-module docstring examples.
